renamed some source files and adjusted imports
closes #461
