+++ fe/src/lib/components/layout/medium/ThemeEditor.svelte
@@ -9,7 +9,7 @@
 		_themeName,
 		_themeNames,
 		_themeVars
-	} from '$lib/stores/theme';
+	} from '$lib/stores/theme.js';
 	import {getHexColor, getThemeClassDefsText} from '$lib/utils/svizzle/style';
 
 	const getVarNames = _.pipe([
