+++ fe/src/lib/components/layout/small/NavSmall.svelte
@@ -14,7 +14,7 @@
 	} from '@svizzle/ui';
 
 	import {makeOnKeyDown} from '$lib/components/svizzle/ui/handlers.js';
-	import {changelogUrl, LOGOS} from '$lib/config';
+	import {changelogUrl, LOGOS} from '$lib/config/app.js';
 	import {
 		_a11yIconFillColor,
 		_a11yIconStrokeColor,
@@ -24,8 +24,8 @@
 		_extLinkTheme,
 		_themeName,
 		toggleTheme,
-	} from '$lib/stores/theme';
-	import {version} from '$lib/utils/version';
+	} from '$lib/stores/theme.js';
+	import {version} from '$lib/utils/version.js';
 
 	export let contentHeight;
 	export let segment;
