+++ fe/src/lib/config/app.js
@@ -1,0 +1,28 @@
+/* tool */
+
+export const toolName = 'Nesta HPMT';
+export const toolLongName = 'Nesta Heat Pump Market Tracker';
+export const contactEmail = 'dataanalytics@nesta.org.uk';
+export const changelogUrl = 'https://github.com/nestauk/asf_hp_market_tracker/blob/staging/CHANGELOG.md';
+
+/* logos */
+
+export const LOGOS = {
+	themeLight: {
+		nesta: '/logos/Nesta-light.svg',
+	},
+	themeDark: {
+		nesta: '/logos/Nesta-dark.svg',
+	}
+}
+
+/* feedback forms */
+
+// survey
+// source: https://docs.google.com/forms/d/1PbnHV3HCqy19aL-05HmZQNdKx5tnUfYrgpH2LKvbmok/edit
+export const surveyFormUrl = 'https://docs.google.com/forms/d/e/1FAIpQLSdavNF1dIsBnwZh_8G1gnNH9R0aTiZTH5Pls_axK4WMkQyhuQ/viewform?embedded=true';
+
+// survey 2
+// source: FIXME add if needed
+// FIXME `survey2FormUrl` unused, same as surveyFormUrl for now
+export const survey2FormUrl = 'https://docs.google.com/forms/d/e/1FAIpQLSdavNF1dIsBnwZh_8G1gnNH9R0aTiZTH5Pls_axK4WMkQyhuQ/viewform?embedded=true'; // FIXME update
