+++ fe/src/lib/config/fonts.js
@@ -1,0 +1,74 @@
+export const fontsInfo = [
+	{
+		family: 'Avenir Next Variable',
+		faces: [
+			{
+				src: 'url(/font/AvenirNext/Variable.ttf) format("truetype")'
+			}
+		]
+	},
+	{
+		family: 'Archivo',
+		faces: [
+			{
+				src: 'url(/font/Archivo/VariableFont_wdth,wght.ttf) format("truetype")',
+				descriptors: {
+					style: 'normal'
+				}
+			},
+			{
+				src: 'url(/font/Archivo/Italic-VariableFont_wdth,wght.ttf) format("truetype")',
+				descriptors: {
+					style: 'italic'
+				}
+			},
+		]
+	},
+	{
+		family: 'Noboto Flex',
+		faces: [
+			{
+				src: 'url(/font/NobotoFlex/Variable.woff2)',
+				descriptors: {
+					weight: 140
+				}
+			}
+		]
+	},
+	{
+		family: 'Courier New'
+	},
+	{
+		family: 'Open Dyslexia',
+		faces: [
+			{
+				src: 'url(/font/OpenDyslexic/Regular.otf) format("opentype")',
+				descriptors: {
+					weight: 400,
+					style: 'normal'
+				}
+			},
+			{
+				src: 'url(/font/OpenDyslexic/Italic.otf) format("opentype")',
+				descriptors: {
+					weight: 400,
+					style: 'italic'
+				}
+			},
+			{
+				src: 'url(/font/OpenDyslexic/Bold.otf) format("opentype")',
+				descriptors: {
+					weight: 700,
+					style: 'normal'
+				}
+			},
+			{
+				src: 'url(/font/OpenDyslexic/BoldItalic.otf) format("opentype")',
+				descriptors: {
+					weight: 700,
+					style: 'italic'
+				}
+			}
+		]
+	}
+];
