+++ fe/src/lib/config/text.js
@@ -1,1 +1,5 @@
 export const noDataMessage = 'No results for the current selection';
+
+/* banners */
+
+export const bannersDefaultFooterText = 'Click on background to dismiss';
