+++ fe/src/routes/explorer/category/geo/[slug]/+page.svelte
@@ -2,7 +2,7 @@
 	import {getKey} from '@svizzle/utils';
 
 	import {page as _page} from '$app/stores';
-	import CatGeoView2 from '$lib/components/explorer/CatGeoView2.svelte';
+	import CatGeoView from '$lib/components/explorer/CatGeoView.svelte';
 	import {interpolateColor} from '$lib/config/colors.js';
 	import {_currentMetric} from '$lib/stores/navigation.js';
 	import {_isViewReady, _viewData} from '$lib/stores/view.js';
@@ -26,7 +26,7 @@
 	}
 </script>
 
-<CatGeoView2
+<CatGeoView
 	{interpolateColor}
 	{items}
 	{keyAccessor}
