+++ fe/src/routes/feedback/+layout.svelte
@@ -4,7 +4,7 @@
 	import * as _ from 'lamb';
 
 	import {page as _page} from '$app/stores';
-	import {_currThemeVars, _linkTheme0} from '$lib/stores/theme';
+	import {_currThemeVars, _linkTheme0} from '$lib/stores/theme.js';
 
 	const segments = ['survey', 'survey2'];
 	const titles = {
