+++ fe/src/routes/feedback/survey/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	import {toolName, surveyFormUrl} from '$lib/config';
+	import {toolName, surveyFormUrl} from '$lib/config/app.js';
 </script>
 
 <svelte:head>
