+++ fe/src/routes/guides/a11ymenu/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	import {toolName} from '$lib/config';
+	import {toolName} from '$lib/config/app.js';
 
 	import A11ymenu from '$lib/_content/guides/A11ymenu.svx';
 </script>
