+++ fe/src/routes/methodology/+layout.svelte
@@ -11,7 +11,7 @@
 	import * as _ from 'lamb';
 
 	import {page as _page} from '$app/stores';
-	import {_currThemeVars, _linkTheme0} from '$lib/stores/theme';
+	import {_currThemeVars, _linkTheme0} from '$lib/stores/theme.js';
 
 	const segments = ['dataProcessing', 'geography'];
 	const titles = {
